+++ clipp/header_modifiers.cpp
@@ -24,8 +24,10 @@
 
 #include "header_modifiers.hpp"
 
+#include <boost/bind.hpp>
 #include <boost/make_shared.hpp>
 #include <boost/foreach.hpp>
+#include <boost/function.hpp>
 #include <boost/algorithm/string/predicate.hpp>
 
 using namespace std;
@@ -42,10 +44,12 @@ class SetModifierDelegate :
 public:
     SetModifierDelegate(
         SetModifier::which_e which,
+        SetModifier::mode_e  mode,
         const string&        key,
         const string&        value
     ) :
         m_which(which),
+        m_mode(mode),
         m_key(key),
         m_value(value)
     {
@@ -67,7 +71,10 @@ public:
     }
 
 private:
-    void modify_header(Input::HeaderEvent& event)
+    void each_matching(
+        Input::HeaderEvent& event,
+        boost::function<bool(Input::header_t&)> f
+    )
     {
         BOOST_FOREACH(Input::header_t& header, event.headers) {
             if (
@@ -80,12 +87,55 @@ private:
                     )
                 )
             ) {
-                header.second = Input::Buffer(m_value);
+                if (! f(header)) {
+                    break;
+                }
             }
         }
     }
 
+    bool set_header(Input::header_t& header)
+    {
+        header.second = Input::Buffer(m_value);
+        return true;
+    }
+
+    static bool is_header(bool& result, Input::header_t& header)
+    {
+        result = true;
+        return false;
+    }
+
+    bool has_header(Input::HeaderEvent& event)
+    {
+        bool result = false;
+        each_matching(event, boost::bind(is_header, boost::ref(result), _1));
+        return result;
+    }
+
+    void modify_header(Input::HeaderEvent& event)
+    {
+        if (m_mode == SetModifier::REPLACE_EXISTING) {
+            each_matching(
+                event,
+                boost::bind(&SetModifierDelegate::set_header, this, _1)
+            );
+        }
+        else if (
+            m_mode == SetModifier::ADD ||
+            (m_mode == SetModifier::ADD_MISSING && ! has_header(event))
+        ) {
+            event.headers.push_back(
+                Input::header_t(
+                    Input::Buffer(m_key),
+                    Input::Buffer(m_value)
+                )
+            );
+        }
+    }
+
     SetModifier::which_e m_which;
+    SetModifier::mode_e m_mode;
     const string& m_key;
     const string& m_value;
 };
@@ -95,18 +145,21 @@ private:
 struct SetModifier::State
 {
     which_e which;
+    mode_e  mode;
     string  key;
     string  value;
 };
 
 SetModifier::SetModifier(
     which_e       which,
+    mode_e        mode,
     const string& key,
     const string& value
 ) :
     m_state(make_shared<State>())
 {
     m_state->which = which;
+    m_state->mode  = mode;
     m_state->key   = key;
     m_state->value = value;
 }
@@ -119,6 +172,7 @@ bool SetModifier::operator()(Input::input_p& input)
 
     SetModifierDelegate delegate(
         m_state->which,
+        m_state->mode,
         m_state->key,
         m_state->value
     );